Extract block narration as util method
