+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/features/ReadCrosshair.java
@@ -4,29 +4,19 @@ import com.github.khanshoaib3.minecraft_access.MainClass;
 import com.github.khanshoaib3.minecraft_access.config.config_maps.RCPartialSpeakingConfigMap;
 import com.github.khanshoaib3.minecraft_access.config.config_maps.RCRelativePositionSoundCueConfigMap;
 import com.github.khanshoaib3.minecraft_access.config.config_maps.ReadCrosshairConfigMap;
-import com.github.khanshoaib3.minecraft_access.mixin.MobSpawnerLogicAccessor;
 import com.github.khanshoaib3.minecraft_access.utils.NarrationUtils;
 import com.github.khanshoaib3.minecraft_access.utils.PlayerUtils;
 import com.github.khanshoaib3.minecraft_access.utils.WorldUtils;
 import com.github.khanshoaib3.minecraft_access.utils.condition.Interval;
-import com.github.khanshoaib3.minecraft_access.utils.position.PlayerPositionUtils;
 import net.minecraft.block.*;
-import net.minecraft.block.entity.BeehiveBlockEntity;
-import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.block.entity.MobSpawnerBlockEntity;
-import net.minecraft.block.entity.SignBlockEntity;
-import net.minecraft.block.enums.ComparatorMode;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.resource.language.I18n;
-import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.fluid.Fluid;
 import net.minecraft.fluid.FluidState;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.entry.RegistryEntry;
-import net.minecraft.registry.tag.BlockTags;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Pair;
@@ -37,15 +27,11 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.Vec3i;
-import org.apache.logging.log4j.util.Strings;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 /**
  * This feature reads the name of the targeted block or entity.<br>
@@ -54,14 +40,6 @@ import java.util.stream.Collectors;
 public class ReadCrosshair {
     private static final double RAY_CAST_DISTANCE = 6.0;
     private static ReadCrosshair instance;
-
-    /**
-     * One redstone wire must be connected with another wire at one of three positions: [side, side down, side up],
-     * since we are checking if the wire is connecting with ALL directions, only take one sample position (x+1) is enough.
-     */
-    public static final Set<Vec3i> THREE_SAMPLE_POSITIONS = Set.of(new Vec3i(1, 0, 0), new Vec3i(1, -1, 0), new Vec3i(1, 1, 0));
-    public static final Predicate<BlockState> IS_REDSTONE_WIRE = (BlockState state) -> state.getBlock() instanceof RedstoneWireBlock;
-
     private boolean enabled;
     private String previousQuery;
     private boolean speakSide;
@@ -213,300 +191,26 @@ public class ReadCrosshair {
         }
 
         BlockPos blockPos = hit.getBlockPos();
-        Pair<String, String> toSpeakAndCurrentQuery = describeBlock(blockPos, side);
-        speakIfFocusChanged(toSpeakAndCurrentQuery.getRight(), toSpeakAndCurrentQuery.getLeft(), Vec3d.of(blockPos));
-    }
+        WorldUtils.BlockInfo blockInfo = WorldUtils.getBlockInfo(blockPos).orElseThrow();
+        // In Minecraft resource location format, for example, "oak_door" for Oak Door.
+        // ref: https://minecraft.wiki/w/Java_Edition_data_values#Blocks)
+        Identifier blockId = Registries.BLOCK.getId(blockInfo.type());
 
-    /**
-     * @param pos  block position (in the client world)
-     * @param side if side is provided, then the invoker is ReadCrosshair
-     * @return toSpeak, currentQuery
-     */
-    public Pair<String, String> describeBlock(BlockPos pos, String side) {
-        MinecraftClient client = MinecraftClient.getInstance();
-        if (Objects.isNull(client)) return new Pair<>("", "");
-        ClientWorld clientWorld = client.world;
-        if (clientWorld == null) return new Pair<>("", "");
-
-        // Since Minecraft uses flyweight pattern for blocks and entities,
-        // All same type of blocks share one singleton Block instance,
-        // While every block keep their states with a BlockState instance.
-        WorldUtils.BlockInfo blockInfo = WorldUtils.getBlockInfo(pos).orElseThrow();
-        BlockPos blockPos = blockInfo.pos();
-        BlockState blockState = blockInfo.state();
-        Block block = blockInfo.type();
-        BlockEntity blockEntity = blockInfo.entity();
-
-        if (enablePartialSpeaking && partialSpeakingBlock) {
-            if (checkIfPartialSpeakingFeatureDoesNotAllowsSpeakingThis(Registries.BLOCK.getId(block))) return new Pair<>("", "");
+        Pair<String, String> toSpeakAndCurrentQuery;
+        if (enablePartialSpeaking && partialSpeakingBlock && checkIfPartialSpeakingFeatureDoesNotAllowsSpeakingThis(blockId)) {
+            toSpeakAndCurrentQuery = new Pair<>("", "");
+        } else {
+            toSpeakAndCurrentQuery = NarrationUtils.narrateBlock(blockPos, side);
         }
 
-        // Difference between toSpeak and currentQuery:
-        // currentQuery is used for checking condition, toSpeak is actually the one to be spoken.
-        // currentQuery is checked to not speak the same block repeatedly, two blocks can have same name.
-        String name = block.getName().getString();
-        String toSpeak = Strings.isBlank(side) ? name : name + " " + side;
-        String currentQuery = name + side;
-
-        // If this config is enabled, add position info to currentQuery,
+        String currentQuery = toSpeakAndCurrentQuery.getRight();
+        String blockPosInString = blockPos.toString();
+        // If "speakingConsecutiveBlocks" config is enabled, add position info to currentQuery,
         // so same blocks at different positions will be regard as different one then trigger the narrator.
         // Class name in production environment can be different
-        String blockPosInString = blockPos.toString();
         if (this.speakingConsecutiveBlocks) currentQuery += " " + blockPosInString;
 
-        // Different special narration (toSpeak) about different type of blocks
-        try {
-            if (blockEntity != null) {
-                // The all signs tag include all types of signs, so it should also work with the hanging signs in 1.20.x
-                if (blockState.isIn(BlockTags.ALL_SIGNS)) {
-                    toSpeak = getSignInfo((SignBlockEntity) blockEntity, client.player, toSpeak);
-                } else if (blockEntity instanceof BeehiveBlockEntity beehiveBlockEntity) {
-                    Pair<String, String> beehiveInfo = getBeehiveInfo(beehiveBlockEntity, blockState, toSpeak, currentQuery);
-                    toSpeak = beehiveInfo.getLeft();
-                    currentQuery = beehiveInfo.getRight();
-                }
-            }
-
-            if (block instanceof PlantBlock || block instanceof CocoaBlock) {
-                Pair<String, String> cropsInfo = getCropsInfo(block, blockState, toSpeak, currentQuery);
-                toSpeak = cropsInfo.getLeft();
-                currentQuery = cropsInfo.getRight();
-            }
-
-            // Check if farmland is wet
-            if (block instanceof FarmlandBlock && blockState.get(FarmlandBlock.MOISTURE) == FarmlandBlock.MAX_MOISTURE) {
-                toSpeak = I18n.translate("minecraft_access.crop.wet_farmland", toSpeak);
-                currentQuery = I18n.translate("minecraft_access.crop.wet_farmland", currentQuery);
-            }
-
-            // Speak monster spawner mob type
-            if (blockEntity instanceof MobSpawnerBlockEntity spawner) {
-                // Will not support non-vanilla custom configured multiple-mob spawner (like generated with command)
-                Entity entity = ((MobSpawnerLogicAccessor) spawner.getLogic()).getRenderedEntity();
-                // Monster spawners that gotten from creative creating screen is empty.
-                String entityName = "Empty";
-                if (entity != null) {
-                    entityName = entity.getDisplayName().getString();
-                }
-                toSpeak = entityName + " " + toSpeak;
-                currentQuery = entityName + currentQuery;
-            }
-
-            // Redstone related
-            Pair<String, String> redstoneRelatedInfo = getRedstoneRelatedInfo(clientWorld, blockPos, block, blockState, toSpeak, currentQuery);
-            toSpeak = redstoneRelatedInfo.getLeft();
-            currentQuery = redstoneRelatedInfo.getRight();
-
-        } catch (Exception e) {
-            MainClass.errorLog("An error occurred while adding narration text for special blocks", e);
-        }
-
-        return new Pair<>(toSpeak, currentQuery);
-    }
-
-    private static String getSignInfo(SignBlockEntity signEntity, ClientPlayerEntity player, String toSpeak) {
-        String[] lines = new String[4];
-
-        for (int i = 0; i < 4; i++) {
-//            lines[i] = signEntity.getTextOnRow(i, false).getString(); // Pre 1.20.x
-            lines[i] = signEntity.getText(signEntity.isPlayerFacingFront(player)).getMessage(i, false).getString();
-        }
-        String content = String.join(", ", lines);
-        return I18n.translate("minecraft_access.read_crosshair.sign_" + (signEntity.isPlayerFacingFront(player) ? "front" : "back") + "_content", toSpeak, content);
-    }
-
-    private static @NotNull Pair<String, String> getRedstoneRelatedInfo(ClientWorld world, BlockPos blockPos, Block block, BlockState blockState, String toSpeak, String currentQuery) {
-        boolean isEmittingPower = world.isEmittingRedstonePower(blockPos, Direction.DOWN);
-        boolean isReceivingPower = world.isReceivingRedstonePower(blockPos);
-
-        if (block instanceof PistonBlock) {
-            String facing = blockState.get(PistonBlock.FACING).getName();
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.facing", toSpeak, I18n.translate("minecraft_access.direction." + facing));
-            currentQuery += "facing " + facing;
-            if (isReceivingPower) {
-                toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-                currentQuery += "powered";
-            }
-        } else if ((block instanceof GlowLichenBlock || block instanceof RedstoneLampBlock) && (isReceivingPower || isEmittingPower)) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-            currentQuery += "powered";
-//        } else if ((block instanceof RedstoneTorchBlock || block instanceof LeverBlock || block instanceof AbstractButtonBlock) && isEmittingPower) { // pre 1.19.3
-        } else if (block instanceof RedstoneWireBlock) {
-            Pair<String, String> p = getRedstoneWireInfo(blockState, blockPos, toSpeak, currentQuery);
-            toSpeak = p.getLeft();
-            currentQuery = p.getRight();
-        } else if ((block instanceof RedstoneTorchBlock || block instanceof LeverBlock || block instanceof ButtonBlock) && isEmittingPower) { // From 1.19.3
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-            currentQuery += "powered";
-        } else if (block instanceof DoorBlock doorBlock && doorBlock.isOpen(blockState)) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.opened", toSpeak);
-            currentQuery += "open";
-        } else if (block instanceof HopperBlock) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.facing", toSpeak, I18n.translate("minecraft_access.direction." + blockState.get(HopperBlock.FACING).getName()));
-            currentQuery += "facing " + blockState.get(HopperBlock.FACING).getName();
-            if (isReceivingPower) {
-                toSpeak = I18n.translate("minecraft_access.read_crosshair.locked", toSpeak);
-                currentQuery += "locked";
-            }
-        } else if (block instanceof ObserverBlock) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.facing", toSpeak, I18n.translate("minecraft_access.direction." + blockState.get(ObserverBlock.FACING).getName()));
-            currentQuery += "facing " + blockState.get(ObserverBlock.FACING).getName();
-            if (isEmittingPower) {
-                toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-                currentQuery += "powered";
-            }
-        } else if (block instanceof DispenserBlock) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.facing", toSpeak, I18n.translate("minecraft_access.direction." + blockState.get(DispenserBlock.FACING).getName()));
-            currentQuery += "facing " + blockState.get(DispenserBlock.FACING).getName();
-            if (isReceivingPower) {
-                toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-                currentQuery += "powered";
-            }
-        } else if (block instanceof ComparatorBlock) {
-            ComparatorMode mode = blockState.get(ComparatorBlock.MODE);
-            Direction facing = blockState.get(ComparatorBlock.FACING);
-            String correctFacing = I18n.translate("minecraft_access.direction." + PlayerPositionUtils.getOppositeDirectionKey(facing.getName()).toLowerCase());
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.comparator_info", toSpeak, correctFacing, mode);
-            if (isReceivingPower) {
-                toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-                currentQuery += "powered";
-            }
-            currentQuery += "mode:" + mode + " facing:" + correctFacing;
-        } else if (block instanceof RepeaterBlock) {
-            boolean locked = blockState.get(RepeaterBlock.LOCKED);
-            int delay = blockState.get(RepeaterBlock.DELAY);
-            Direction facing = blockState.get(ComparatorBlock.FACING);
-            String correctFacing = I18n.translate("minecraft_access.direction." + PlayerPositionUtils.getOppositeDirectionKey(facing.getName()).toLowerCase());
-
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.repeater_info", toSpeak, correctFacing, delay);
-            currentQuery += "delay:" + delay + " facing:" + correctFacing;
-            if (locked) {
-                toSpeak = I18n.translate("minecraft_access.read_crosshair.locked", toSpeak);
-                currentQuery += "locked";
-            }
-        } else if (isReceivingPower) { // For all the other blocks
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.powered", toSpeak);
-            currentQuery += "powered";
-        }
-
-        return new Pair<>(toSpeak, currentQuery);
-    }
-
-    private static @NotNull Pair<String, String> getRedstoneWireInfo(BlockState blockState, BlockPos pos, String toSpeak, String currentQuery) {
-        int powerLevel = blockState.get(RedstoneWireBlock.POWER);
-        if (powerLevel > 0) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.redstone_wire_power", toSpeak, powerLevel);
-            currentQuery += "power level " + powerLevel;
-        }
-
-        List<String> connectedDirections = Direction.Type.HORIZONTAL.stream()
-                .map(direction -> {
-                    String directionName = I18n.translate("minecraft_access.direction." + direction.getName());
-
-                    switch (blockState.get(RedstoneWireBlock.DIRECTION_TO_WIRE_CONNECTION_PROPERTY.get(direction))) {
-                        case UP -> {
-                            return directionName + " " + I18n.translate("minecraft_access.direction.up");
-                        }
-                        case SIDE -> {
-                            return directionName;
-                        }
-                        default -> {
-                            return null;
-                        }
-                    }
-                })
-                .filter(Objects::nonNull)
-                .collect(Collectors.toList());
-
-        // Unconnected redstone dust now has all direction block states set to "side" since 20w18a (before 1.16)
-        // https://minecraft.wiki/w/Redstone_Dust
-        // So here is an additional check to see if the redstone wire is really connected to all directions
-        if (connectedDirections.size() == 4) {
-            Optional<Boolean> result = WorldUtils.checkAnyOfSurroundingBlocks(pos, THREE_SAMPLE_POSITIONS, IS_REDSTONE_WIRE);
-            if (result.isEmpty()) return new Pair<>(toSpeak, currentQuery);
-            // return early if we find out that this wire is not connected to all directions
-            if (Boolean.FALSE.equals(result.get())) return new Pair<>(toSpeak, currentQuery);
-        }
-
-        String directionsToSpeak = String.join(I18n.translate("minecraft_access.other.words_connection"), connectedDirections);
-        toSpeak = I18n.translate("minecraft_access.read_crosshair.redstone_wire_connection", toSpeak, directionsToSpeak);
-        currentQuery += "connected to " + connectedDirections;
-
-        return new Pair<>(toSpeak, currentQuery);
-    }
-
-    private static @NotNull Pair<String, String> getBeehiveInfo(BeehiveBlockEntity blockEntity, BlockState blockState, String toSpeak, String currentQuery) {
-        boolean isSmoked = blockEntity.isSmoked();
-        int honeyLevel = blockState.get(BeehiveBlock.HONEY_LEVEL);
-        Direction facingDirection = blockState.get(BeehiveBlock.FACING);
-
-        if (isSmoked) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.bee_hive_smoked", toSpeak);
-            currentQuery += "smoked";
-        }
-
-        if (honeyLevel > 0) {
-            toSpeak = I18n.translate("minecraft_access.read_crosshair.bee_hive_honey_level", toSpeak, honeyLevel);
-            currentQuery += ("honey-level:" + honeyLevel);
-        }
-
-        toSpeak = I18n.translate("minecraft_access.read_crosshair.bee_hive_facing", toSpeak, facingDirection.getName());
-        currentQuery += ("facing:" + facingDirection.getName());
-
-        return new Pair<>(toSpeak, currentQuery);
-    }
-
-    /**
-     * Blocks that can be planted and have growing stages (age) and harvestable.<br>
-     * Including wheat, carrot, potato, beetroot, nether wart, cocoa bean,
-     * torch flower, pitcher crop.<br>
-     * Watermelon vein and pumpkin vein are not harvestable so not be included here.
-     */
-    private static @NotNull Pair<String, String> getCropsInfo(Block block, BlockState blockState, String toSpeak, String currentQuery) {
-        int currentAge, maxAge;
-
-        if (block instanceof CropBlock) {
-            if (block instanceof BeetrootsBlock) {
-                // Beetroots has a different max_age as 3
-                currentAge = blockState.get(BeetrootsBlock.AGE);
-                maxAge = BeetrootsBlock.BEETROOTS_MAX_AGE;
-            } else if (block instanceof TorchflowerBlock) {
-                currentAge = blockState.get(TorchflowerBlock.AGE);
-                maxAge = 2;
-            } else {
-                // While wheat, carrots, potatoes has max_age as 7
-                currentAge = blockState.get(CropBlock.AGE);
-                maxAge = CropBlock.MAX_AGE;
-            }
-        } else if (block instanceof CocoaBlock) {
-            currentAge = blockState.get(CocoaBlock.AGE);
-            maxAge = CocoaBlock.MAX_AGE;
-        } else if (block instanceof NetherWartBlock) {
-            currentAge = blockState.get(NetherWartBlock.AGE);
-            // The max_age of NetherWartBlock hasn't been translated, for future compatibility, hard code it.
-            maxAge = 3;
-        } else if (block instanceof PitcherCropBlock) {
-            currentAge = blockState.get(PitcherCropBlock.AGE);
-            maxAge = 4;
-        } else {
-            return new Pair<>(toSpeak, currentQuery);
-        }
-
-        String configKey = checkCropRipeLevel(currentAge, maxAge);
-        return new Pair<>(I18n.translate(configKey, toSpeak), I18n.translate(configKey, currentQuery));
-    }
-
-    /**
-     * @return corresponding ripe level text config key
-     */
-    private static String checkCropRipeLevel(Integer current, int max) {
-        if (current >= max) {
-            return "minecraft_access.crop.ripe";
-        } else if (current < max / 2) {
-            return "minecraft_access.crop.seedling";
-        } else {
-            return "minecraft_access.crop.half_ripe";
-        }
+        speakIfFocusChanged(currentQuery, toSpeakAndCurrentQuery.getLeft(), Vec3d.of(blockPos));
     }
 
     private boolean checkForFluidHit(MinecraftClient minecraftClient, HitResult fluidHit) {
